pyproject.toml
[tool.poetry]
name = "google-smart-home-integration"
version = "0.1.0"
description = ""
authors = ["Brenno Almeida <contato@brennoflavio.com.br>"]
readme = "README.md"
packages = [{include = "google_smart_home_integration"}]

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "2.111.0"
google-auth-oauthlib = "1.2.0"
oauth2client = "4.1.3"
caldav = "0.11.0"
python-dateutil = "2.8.2"
flask = "3.0.0"
ics = "0.7.2"
flask-httpauth = "4.8.0"
gkeepapi = "0.15.1"
requests = "2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
